+++ CHANGELOG.md
@@ -28,6 +28,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Update S3 deployment documentation to include CORS configuration note - #3743 by @NyanKiyoshi
 - Fix missing migrations for is_published field of product and page model - #3757 by @jxltom
 - Add header and footer for checkout success pages #3752 by @jxltom
+- Filter order by payment status from order's last payment - #3749 @jxltom
 
 
 ## 2.3.0